Lazily retrieve the type methods
Previously invocableFromLink would look for the type and then ask
the type to do the dirty work. This leads to a lot of web requests
though, if the class inherits quite a few methods. An example for
this is "javax.swing.JButton#getText(" which would crash the bot.

This commit just never asks the type and resolves it itself.